WIP: address configs
