+++ src/main/java/com/api/bigu/BiguApplication.java
@@ -1,12 +1,11 @@
 package com.api.bigu;
 
-import com.api.bigu.auth.AuthenticationService;
-import com.api.bigu.auth.RegisterRequest;
+import com.api.bigu.services.AuthenticationService;
+import com.api.bigu.dto.auth.RegisterRequest;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 import static com.api.bigu.models.enums.Role.ADMIN;
 import static com.api.bigu.models.enums.Role.USER;
