--- src/main/java/com/api/bigu/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package com.api.bigu.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-//@Configuration
-//@EnableSwagger2
-public class SwaggerConfig {
-
-//    @Bean
-//    public Docket api() {
-//        return new Docket(DocumentationType.SWAGGER_2)
-//                .select()
-//                .apis(RequestHandlerSelectors.basePackage("com.api.bigu"))
-//                .paths(PathSelectors.any())
-//                .build();
-//    }
-
-}
