+++ src/main/java/com/api/bigu/controllers/AddressController.java
@@ -1,0 +1,12 @@
+package com.api.bigu.controllers;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "/address")
+public class AddressController {
+
+
+
+}
