+++ src/main/java/com/api/bigu/dto/address/AddressRequest.java
@@ -1,0 +1,4 @@
+package com.api.bigu.dto.address;
+
+public class AddressRequest {
+}
