+++ src/main/java/com/api/bigu/dto/auth/AuthenticationRequest.java
@@ -1,7 +1,5 @@
-package com.api.bigu.auth;
+package com.api.bigu.dto.auth;
 
-import com.api.bigu.models.enums.Role;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,13 +10,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class RegisterRequest {
-
-    private String fullName;
+public class AuthenticationRequest {
 
     @Pattern(regexp = "[\\w-.]+@([\\w-])+.ufcg.edu.+[\\w-]$", message = "email not valid")
     private String email;
-    private String phoneNumber;
     private String password;
-    private Role role;
 }
