+++ src/main/java/com/api/bigu/dto/auth/RegisterRequest.java
@@ -1,5 +1,7 @@
-package com.api.bigu.auth;
+package com.api.bigu.dto.auth;
 
+import com.api.bigu.models.enums.Role;
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,9 +12,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
+public class RegisterRequest {
 
+    private String fullName;
     @Pattern(regexp = "[\\w-.]+@([\\w-])+.ufcg.edu.+[\\w-]$", message = "email not valid")
     private String email;
+    private String phoneNumber;
     private String password;
+    private Role role;
 }
